requirements.txt
fastapi~=0.105.0
uvicorn
langchain
langchain_community
langchain-openai
python-dotenv~=1.0.0
python-multipart
pydantic~=2.9.2
pypdf
sentence-transformers
weasyprint
langchain_huggingface
bs4
requests
faiss-cpu
react-icons
beautifulsoup4==4.10.0
weasyprint==53.3
openai~=1.53.0